+++ src/internet/model/tcp-socket-state.cc
@@ -91,7 +91,11 @@ TcpSocketState::GetTypeId()
                             MakeTraceSourceAccessor(&TcpSocketState::m_bytesInFlight),
                             "ns3::TracedValueCallback::Uint32")
             .AddTraceSource("RTT",
-                            "Last RTT sample",
+                            "Smoothed RTT",
+                            MakeTraceSourceAccessor(&TcpSocketState::m_srtt),
+                            "ns3::TracedValueCallback::Time")
+            .AddTraceSource("LastRTT",
+                            "RTT of the last (S)ACKed packet",
                             MakeTraceSourceAccessor(&TcpSocketState::m_lastRtt),
                             "ns3::TracedValueCallback::Time");
     return tid;
@@ -120,6 +124,7 @@ TcpSocketState::TcpSocketState(const TcpSocketState& other)
       m_minRtt(other.m_minRtt),
       m_bytesInFlight(other.m_bytesInFlight),
       m_isCwndLimited(other.m_isCwndLimited),
+      m_srtt(other.m_srtt),
       m_lastRtt(other.m_lastRtt),
       m_ecnMode(other.m_ecnMode),
       m_useEcn(other.m_useEcn),